+++ pkg/httpx/httpx.go
@@ -3,5 +3,7 @@ package httpx
 const (
 	HeaderContentType   = "Content-Type"
 	HeaderContentLength = "Content-Length"
+	HeaderContentRange  = "Content-Range"
+	HeaderRange         = "Range"
 	HeaderAcceptRanges  = "Accept-Ranges"
 )