+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [#905](https://github.com/spegel-org/spegel/pull/905) Change mirror type to url and add byte range parameter.
+
 ### Changed
 
 ### Deprecated